+++ src/commandGroups/all.ts
@@ -4,10 +4,12 @@ import { commandTable as otherShells } from "./othershells.js";
 import { commandTable as processCmds } from "./processCommands.js";
 import { commandTable as streamsC } from "./streams.js";
 import { commandTable as thatShell } from "./thatShell.js";
+import { commandTable as thatShellUserFriendly } from "./thatShellUserFriendly.js";
 import { commandTable as webGroup } from "./webGroup.js";
 import { commandTable as uptime} from "../commands/uptime.js";
 import { commandTable as meow} from "../commands/meow.js";
 import { commandCollection } from "../tools/commandCollection.js";
+
 
 const commandTable = {
     ...pipeGroup,
@@ -15,6 +17,7 @@ const commandTable = {
     ...processCmds,
     ...streamsC,
     ...thatShell,
+    ...thatShellUserFriendly,
     ...webGroup,
     ...uptime,
     ...meow