+++ src/commandGroups/thatShellUserFriendly.ts
@@ -1,0 +1,24 @@
+import { commandTable as logsZ} from "../commands/logs.js";
+import { commandTable as hideZ } from "../commands/hide.js";
+import { commandTable as helpZ } from "../commands/help.js";
+import { commandTable as clearZ } from "../commands/clear.js";
+import { commandTable as bindZ } from "../commands/bind.js";
+import { commandTable as existsx } from "../commands/exists.js";
+import { commandTable as versionZ } from "../commands/version.js";
+import { commandTable as term } from "../commands/terminal.js";
+import { cmdTableToCommandCompounts } from "../tools/commandCreatorTools.js";
+import { commandCollection } from "../tools/commandCollection.js";
+
+
+const commandTable = {
+    ...helpZ,
+    ...clearZ,
+    ...bindZ,
+    ...versionZ
+};
+
+const compounds = cmdTableToCommandCompounts(commandTable);
+
+const collection = new commandCollection(commandTable);
+
+export {commandTable, compounds, collection}