+++ src/tools/quickSetup.ts
@@ -1,3 +1,4 @@
+import { existsSync, mkdirSync } from "fs";
 import { cd } from "../apis/commands/osApis/filesystem.js";
 import { collection } from "../commandGroups/all.js";
 import { configDataProvide, constructConfig } from "../config.js";
@@ -6,6 +7,7 @@ import { setUpInterrupsForProcess } from "../interrup.js";
 import { createNewTerminal } from "../programdata.js";
 import { logSystemError } from "../ultrabasic.js";
 import { welcome } from "./welcome.js";
+import { createTerminalQuick } from "../apis/allApis.js";
 
 
 
@@ -23,33 +25,43 @@ import { welcome } from "./welcome.js";
  * @param configD the config of that (OPTIONAL)
  * @param chwdToSelectedCwd whether to sync cwd
  */
-function quickSetup(configD: configDataProvide = {}, chwdToSelectedCwd: boolean = true){
+function quickSetup(configD: configDataProvide = {}){
     const toSend = {commandTable: collection};
 
     Object.assign(toSend, configD);
 
-    const nT = createNewTerminal(
-        "main", 
-        {
-            config: constructConfig(toSend),
-            out: process.stdout,
-            in: process.stdin,
-        }
-    );
+    const nT = createTerminalQuick("main", {
+        config: toSend,
+        out: process.stdout,
+        in: process.stdin,
+        setupProcessInterrups: true,
+        chwdToSelectedCwd: true
+    });
+
+    // const nT = createNewTerminal(
+    //     "main", 
+    //     {
+    //         config: constructConfig(toSend),
+    //         out: process.stdout,
+    //         in: process.stdin,
+    //     }
+    // );
 
     if(!nT){
         throw new logSystemError("The quick system was not able to create a main terminal properly");
     }
 
-    if(chwdToSelectedCwd && nT.procLinked){
-        nT.procLinked.chdir(nT.cwd);
-    }
+    // if(chwdToSelectedCwd && nT.procLinked){
+    //     if(!existsSync(nT.cwd)){
+    //         mkdirSync(nT.cwd, {recursive: true});
+    //     }
+    //     nT.procLinked.chdir(nT.cwd);
+    // }
 
-    welcome(nT);
+    // welcome(nT);
 
 
-    setupInHandlerListener(nT);
-    setUpInterrupsForProcess(nT);
+    // setupInHandlerListener(nT);
 }
 
 export {quickSetup}